+++ Control/Joint/Abilities.hs
@@ -1,6 +1,6 @@
 module Control.Joint.Abilities (module Exports) where
 
 import Control.Joint.Abilities.Liftable as Exports
-import Control.Joint.Abilities.Adaptable as Exports
+import Control.Joint.Abilities.Completable as Exports
 import Control.Joint.Abilities.Transformer as Exports
 import Control.Joint.Abilities.Interpreted as Exports