+++ Control/Joint/Abilities/Adaptable.hs
@@ -1,6 +1,0 @@
-module Control.Joint.Abilities.Adaptable where
-
-import Control.Joint.Core (type (~>))
-
-class Adaptable (subeff :: * -> *) (eff :: * -> *) | subeff -> eff where
-	adapt :: subeff ~> eff