+++ Control/Joint/Abilities/Completable.hs
@@ -1,0 +1,6 @@
+module Control.Joint.Abilities.Completable where
+
+import Control.Joint.Core (type (~>))
+
+class Completable (subeff :: * -> *) (eff :: * -> *) | subeff -> eff where
+	complete :: subeff ~> eff