+++ CHANGELOG.md
@@ -43,3 +43,4 @@
 # 0.1.7
 * Create `Operators` module to work with nested functor compositions
 * Add experimental `$>>=` and `>>=$` monadic operators
+* Define `replace` method in `Effects.State` module