+++ CHANGELOG.md
@@ -46,3 +46,4 @@
 * Define `replace` method in `Effects.State` module
 * Change order of parameters in `TUT` newtype schema
 * Change arity of `Schema` associated type family
+* Add `transformers` dependency